修正一些代码风格和命名；

+++ app/src/main/java/me/drakeet/inmessage/utils/TimeUtils.java
@@ -60,7 +60,7 @@ public class TimeUtils {
         cal.set(Calendar.SECOND, 0);
         cal.set(Calendar.MINUTE, 0);
         cal.set(Calendar.MILLISECOND, 0);
-        return (long) (cal.getTimeInMillis());
+        return cal.getTimeInMillis();
     }
 
     public static Date getDateFromTimestamp(long longTime) {
